fix: Check DBC version after turning it on when empty
When the DBC has not been started since the last MDB update, the version info
will be empty. In this case, we now turn the DBC on once, wait for readiness,
and then check the version again.

Once we have Redis persistence across updates, this will not be needed anymore.